pyproject.toml
[project]
name = "meridian-compliance-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.3",
    "azure-keyvault-secrets>=4.11.0",
    "langgraph>=1.1.7",
    "langsmith>=0.7.32",
    "opentelemetry-api>=1.41.0",
    "opentelemetry-sdk>=1.41.0",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
]
